+++ example/lib/common/tu_chong_source.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 class TuChongSource {
   int counts;
-  bool is_history;
+  bool isHistory;
   bool more;
   String message;
   String result;
@@ -12,7 +12,7 @@ class TuChongSource {
 
   TuChongSource.fromParams(
       {this.counts,
-      this.is_history,
+      this.isHistory,
       this.more,
       this.message,
       this.result,
@@ -26,7 +26,7 @@ class TuChongSource {
 
   TuChongSource.fromJson(jsonRes) {
     counts = jsonRes['counts'];
-    is_history = jsonRes['is_history'];
+    isHistory = jsonRes['is_history'];
     more = jsonRes['more'];
     message = jsonRes['message'];
     result = jsonRes['result'];
@@ -40,42 +40,44 @@ class TuChongSource {
 
   @override
   String toString() {
-    return '{"counts": $counts,"is_history": $is_history,"more": $more,"message": ${message != null ? '${json.encode(message)}' : 'null'},"result": ${result != null ? '${json.encode(result)}' : 'null'},"feedList": $feedList}';
+    return '{"counts": $counts,"is_history": $isHistory,"more": $more,"message": ${message != null ? '${json.encode(message)}' : 'null'},"result": ${result != null ? '${json.encode(result)}' : 'null'},"feedList": $feedList}';
   }
 }
 
 class TuChongItem {
-  Object title_image;
+  Object titleImage;
   int comments;
   int favorites;
-  int image_count;
-  int post_id;
+  int imageCount;
+  int postId;
+  int shares;
   int views;
   bool collected;
   bool delete;
-  bool is_favorite;
+  bool isFavorite;
+  bool recommend;
   bool rewardable;
   bool update;
-  String author_id;
+  String authorId;
   String content;
-  String created_at;
-  String data_type;
+  String createdAt;
+  String dataType;
   String excerpt;
-  String parent_comments;
-  String passed_time;
-  String published_at;
-  String recom_type;
+  String parentComments;
+  String passedTime;
+  String publishedAt;
+  String recomType;
   String rewards;
-  String rqt_id;
-  String site_id;
+  String rqtId;
+  String siteId;
   String title;
   String type;
   String url;
-  List<dynamic> comment_list_prefix;
-  List<String> event_tags;
-  List<dynamic> favorite_list_prefix;
+  List<dynamic> commentListPrefix;
+  List<String> eventTags;
+  List<dynamic> favoriteListPrefix;
   List<ImageItem> images;
-  List<dynamic> reward_list_prefix;
+  List<dynamic> rewardListPrefix;
   List<dynamic> sites;
   List<String> tags;
   Site site;
@@ -92,95 +94,97 @@ class TuChongItem {
   String get imageUrl {
     if (!hasImage) return "";
     return "https://photo.tuchong.com/" +
-        images[0].user_id.toString() +
+        images[0].userId.toString() +
         "/f/" +
-        images[0].img_id.toString() +
+        images[0].imgId.toString() +
         ".jpg";
   }
 
   TuChongItem.fromParams(
-      {this.title_image,
+      {this.titleImage,
       this.comments,
       this.favorites,
-      this.image_count,
-      this.post_id,
+      this.imageCount,
+      this.postId,
+      this.shares,
       this.views,
       this.collected,
       this.delete,
-      this.is_favorite,
+      this.isFavorite,
+      this.recommend,
       this.rewardable,
       this.update,
-      this.author_id,
+      this.authorId,
       this.content,
-      this.created_at,
-      this.data_type,
+      this.createdAt,
+      this.dataType,
       this.excerpt,
-      this.parent_comments,
-      this.passed_time,
-      this.published_at,
-      this.recom_type,
+      this.parentComments,
+      this.passedTime,
+      this.publishedAt,
+      this.recomType,
       this.rewards,
-      this.rqt_id,
-      this.site_id,
+      this.rqtId,
+      this.siteId,
       this.title,
       this.type,
       this.url,
-      this.comment_list_prefix,
-      this.event_tags,
-      this.favorite_list_prefix,
+      this.commentListPrefix,
+      this.eventTags,
+      this.favoriteListPrefix,
       this.images,
-      this.reward_list_prefix,
+      this.rewardListPrefix,
       this.sites,
       this.tags,
       this.site});
 
   TuChongItem.fromJson(jsonRes) {
-    title_image = jsonRes['title_image'];
+    titleImage = jsonRes['title_image'];
     comments = jsonRes['comments'];
     favorites = jsonRes['favorites'];
-    image_count = jsonRes['image_count'];
-    post_id = jsonRes['post_id'];
+    imageCount = jsonRes['image_count'];
+    postId = jsonRes['post_id'];
+    shares = jsonRes['shares'];
     views = jsonRes['views'];
     collected = jsonRes['collected'];
     delete = jsonRes['delete'];
-    is_favorite = jsonRes['is_favorite'];
+    isFavorite = jsonRes['is_favorite'];
+    recommend = jsonRes['recommend'];
     rewardable = jsonRes['rewardable'];
     update = jsonRes['update'];
-    author_id = jsonRes['author_id'];
+    authorId = jsonRes['author_id'];
     content = jsonRes['content'];
-    created_at = jsonRes['created_at'];
-    data_type = jsonRes['data_type'];
+    createdAt = jsonRes['created_at'];
+    dataType = jsonRes['data_type'];
     excerpt = jsonRes['excerpt'];
-    parent_comments = jsonRes['parent_comments'];
-    passed_time = jsonRes['passed_time'];
-    published_at = jsonRes['published_at'];
-    recom_type = jsonRes['recom_type'];
+    parentComments = jsonRes['parent_comments'];
+    passedTime = jsonRes['passed_time'];
+    publishedAt = jsonRes['published_at'];
+    recomType = jsonRes['recom_type'];
     rewards = jsonRes['rewards'];
-    rqt_id = jsonRes['rqt_id'];
-    site_id = jsonRes['site_id'];
+    rqtId = jsonRes['rqt_id'];
+    siteId = jsonRes['site_id'];
     title = jsonRes['title'];
     type = jsonRes['type'];
     url = jsonRes['url'];
-    comment_list_prefix = jsonRes['comment_list_prefix'] == null ? null : [];
+    commentListPrefix = jsonRes['comment_list_prefix'] == null ? null : [];
 
-    for (var comment_list_prefixItem
-        in comment_list_prefix == null ? [] : jsonRes['comment_list_prefix']) {
-      comment_list_prefix.add(comment_list_prefixItem);
+    for (var commentListPrefixItem
+        in commentListPrefix == null ? [] : jsonRes['comment_list_prefix']) {
+      commentListPrefix.add(commentListPrefixItem);
     }
 
-    event_tags = jsonRes['event_tags'] == null ? null : [];
+    eventTags = jsonRes['event_tags'] == null ? null : [];
 
-    for (var event_tagsItem
-        in event_tags == null ? [] : jsonRes['event_tags']) {
-      event_tags.add(event_tagsItem);
+    for (var eventTagsItem in eventTags == null ? [] : jsonRes['event_tags']) {
+      eventTags.add(eventTagsItem);
     }
 
-    favorite_list_prefix = jsonRes['favorite_list_prefix'] == null ? null : [];
+    favoriteListPrefix = jsonRes['favorite_list_prefix'] == null ? null : [];
 
-    for (var favorite_list_prefixItem in favorite_list_prefix == null
-        ? []
-        : jsonRes['favorite_list_prefix']) {
-      favorite_list_prefix.add(favorite_list_prefixItem);
+    for (var favoriteListPrefixItem
+        in favoriteListPrefix == null ? [] : jsonRes['favorite_list_prefix']) {
+      favoriteListPrefix.add(favoriteListPrefixItem);
     }
 
     images = jsonRes['images'] == null ? null : [];
@@ -190,11 +194,11 @@ class TuChongItem {
           .add(imagesItem == null ? null : new ImageItem.fromJson(imagesItem));
     }
 
-    reward_list_prefix = jsonRes['reward_list_prefix'] == null ? null : [];
+    rewardListPrefix = jsonRes['reward_list_prefix'] == null ? null : [];
 
-    for (var reward_list_prefixItem
-        in reward_list_prefix == null ? [] : jsonRes['reward_list_prefix']) {
-      reward_list_prefix.add(reward_list_prefixItem);
+    for (var rewardListPrefixItem
+        in rewardListPrefix == null ? [] : jsonRes['reward_list_prefix']) {
+      rewardListPrefix.add(rewardListPrefixItem);
     }
 
     sites = jsonRes['sites'] == null ? null : [];
@@ -214,104 +218,86 @@ class TuChongItem {
 
   @override
   String toString() {
-    return '{"title_image": ${title != null ? '${json.encode(title)}' : 'null'}_image,"comments": $comments,"favorites": $favorites,"image_count": $image_count,"post_id": $post_id,"views": $views,"collected": $collected,"delete": $delete,"is_favorite": $is_favorite,"rewardable": $rewardable,"update": $update,"author_id": ${author_id != null ? '${json.encode(author_id)}' : 'null'},"content": ${content != null ? '${json.encode(content)}' : 'null'},"created_at": ${created_at != null ? '${json.encode(created_at)}' : 'null'},"data_type": ${data_type != null ? '${json.encode(data_type)}' : 'null'},"excerpt": ${excerpt != null ? '${json.encode(excerpt)}' : 'null'},"parent_comments": ${parent_comments != null ? '${json.encode(parent_comments)}' : 'null'},"passed_time": ${passed_time != null ? '${json.encode(passed_time)}' : 'null'},"published_at": ${published_at != null ? '${json.encode(published_at)}' : 'null'},"recom_type": ${recom_type != null ? '${json.encode(recom_type)}' : 'null'},"rewards": ${rewards != null ? '${json.encode(rewards)}' : 'null'},"rqt_id": ${rqt_id != null ? '${json.encode(rqt_id)}' : 'null'},"site_id": ${site_id != null ? '${json.encode(site_id)}' : 'null'},"title": ${title != null ? '${json.encode(title)}' : 'null'},"type": ${type != null ? '${json.encode(type)}' : 'null'},"url": ${url != null ? '${json.encode(url)}' : 'null'},"comment_list_prefix": $comment_list_prefix,"event_tags": $event_tags,"favorite_list_prefix": $favorite_list_prefix,"images": $images,"reward_list_prefix": $reward_list_prefix,"sites": $sites,"tags": $tags,"site": $site}';
+    return '{"title_image": ${title != null ? '${json.encode(title)}' : 'null'}Image,"comments": $comments,"favorites": $favorites,"image_count": $imageCount,"post_id": $postId,"shares": $shares,"views": $views,"collected": $collected,"delete": $delete,"is_favorite": $isFavorite,"recommend": $recommend,"rewardable": $rewardable,"update": $update,"author_id": ${authorId != null ? '${json.encode(authorId)}' : 'null'},"content": ${content != null ? '${json.encode(content)}' : 'null'},"created_at": ${createdAt != null ? '${json.encode(createdAt)}' : 'null'},"data_type": ${dataType != null ? '${json.encode(dataType)}' : 'null'},"excerpt": ${excerpt != null ? '${json.encode(excerpt)}' : 'null'},"parent_comments": ${parentComments != null ? '${json.encode(parentComments)}' : 'null'},"passed_time": ${passedTime != null ? '${json.encode(passedTime)}' : 'null'},"published_at": ${publishedAt != null ? '${json.encode(publishedAt)}' : 'null'},"recom_type": ${recomType != null ? '${json.encode(recomType)}' : 'null'},"rewards": ${rewards != null ? '${json.encode(rewards)}' : 'null'},"rqt_id": ${rqtId != null ? '${json.encode(rqtId)}' : 'null'},"site_id": ${siteId != null ? '${json.encode(siteId)}' : 'null'},"title": ${title != null ? '${json.encode(title)}' : 'null'},"type": ${type != null ? '${json.encode(type)}' : 'null'},"url": ${url != null ? '${json.encode(url)}' : 'null'},"comment_list_prefix": $commentListPrefix,"event_tags": $eventTags,"favorite_list_prefix": $favoriteListPrefix,"images": $images,"reward_list_prefix": $rewardListPrefix,"sites": $sites,"tags": $tags,"site": $site}';
   }
-
-  @override
-  bool operator ==(other) {
-    // TODO: implement ==
-    return other.post_id == post_id;
-  }
-
-  @override
-  // TODO: implement hashCode
-  int get hashCode => post_id;
 }
 
 class Site {
   int followers;
   int verifications;
-  int verified_type;
-  bool is_following;
+  int verifiedType;
+  bool hasEverphotoNote;
+  bool isBindEverphoto;
+  bool isFollowing;
   bool verified;
   String description;
   String domain;
   String icon;
   String name;
-  String site_id;
+  String siteId;
   String type;
   String url;
-  String verified_reason;
-  List<VerificationItem> verification_list;
+  String verifiedReason;
+  List<dynamic> verificationList;
 
   Site.fromParams(
       {this.followers,
       this.verifications,
-      this.verified_type,
-      this.is_following,
+      this.verifiedType,
+      this.hasEverphotoNote,
+      this.isBindEverphoto,
+      this.isFollowing,
       this.verified,
       this.description,
       this.domain,
       this.icon,
       this.name,
-      this.site_id,
+      this.siteId,
       this.type,
       this.url,
-      this.verified_reason,
-      this.verification_list});
+      this.verifiedReason,
+      this.verificationList});
 
   Site.fromJson(jsonRes) {
     followers = jsonRes['followers'];
     verifications = jsonRes['verifications'];
-    verified_type = jsonRes['verified_type'];
-    is_following = jsonRes['is_following'];
+    verifiedType = jsonRes['verified_type'];
+    hasEverphotoNote = jsonRes['has_everphoto_note'];
+    isBindEverphoto = jsonRes['is_bind_everphoto'];
+    isFollowing = jsonRes['is_following'];
     verified = jsonRes['verified'];
     description = jsonRes['description'];
     domain = jsonRes['domain'];
     icon = jsonRes['icon'];
     name = jsonRes['name'];
-    site_id = jsonRes['site_id'];
+    var id = jsonRes['site_id'];
+    if (id is int) {
+      siteId = id.toString();
+    } else if (id is String) {
+      siteId = id;
+    }
+
     type = jsonRes['type'];
     url = jsonRes['url'];
-    verified_reason = jsonRes['verified_reason'];
-    verification_list = jsonRes['verification_list'] == null ? null : [];
+    verifiedReason = jsonRes['verified_reason'];
+    verificationList = jsonRes['verification_list'] == null ? null : [];
 
-    for (var verification_listItem
-        in verification_list == null ? [] : jsonRes['verification_list']) {
-      verification_list.add(verification_listItem == null
-          ? null
-          : new VerificationItem.fromJson(verification_listItem));
+    for (var verificationListItem
+        in verificationList == null ? [] : jsonRes['verification_list']) {
+      verificationList.add(verificationListItem);
     }
   }
 
   @override
   String toString() {
-    return '{"followers": $followers,"verifications": $verifications,"verified_type": $verified_type,"is_following": $is_following,"verified": $verified,"description": ${description != null ? '${json.encode(description)}' : 'null'},"domain": ${domain != null ? '${json.encode(domain)}' : 'null'},"icon": ${icon != null ? '${json.encode(icon)}' : 'null'},"name": ${name != null ? '${json.encode(name)}' : 'null'},"site_id": ${site_id != null ? '${json.encode(site_id)}' : 'null'},"type": ${type != null ? '${json.encode(type)}' : 'null'},"url": ${url != null ? '${json.encode(url)}' : 'null'},"verified_reason": ${verified_reason != null ? '${json.encode(verified_reason)}' : 'null'},"verification_list": $verification_list}';
-  }
-}
-
-class VerificationItem {
-  int verification_type;
-  String verification_reason;
-
-  VerificationItem.fromParams(
-      {this.verification_type, this.verification_reason});
-
-  VerificationItem.fromJson(jsonRes) {
-    verification_type = jsonRes['verification_type'];
-    verification_reason = jsonRes['verification_reason'];
-  }
-
-  @override
-  String toString() {
-    return '{"verification_type": $verification_type,"verification_reason": ${verification_reason != null ? '${json.encode(verification_reason)}' : 'null'}}';
+    return '{"followers": $followers,"verifications": $verifications,"verified_type": $verifiedType,"has_everphoto_note": $hasEverphotoNote,"is_bind_everphoto": $isBindEverphoto,"is_following": $isFollowing,"verified": $verified,"description": ${description != null ? '${json.encode(description)}' : 'null'},"domain": ${domain != null ? '${json.encode(domain)}' : 'null'},"icon": ${icon != null ? '${json.encode(icon)}' : 'null'},"name": ${name != null ? '${json.encode(name)}' : 'null'},"site_id": ${siteId != null ? '${json.encode(siteId)}' : 'null'},"type": ${type != null ? '${json.encode(type)}' : 'null'},"url": ${url != null ? '${json.encode(url)}' : 'null'},"verified_reason": ${verifiedReason != null ? '${json.encode(verifiedReason)}' : 'null'},"verification_list": $verificationList}';
   }
 }
 
 class ImageItem {
   int height;
-  int img_id;
-  int user_id;
+  int imgId;
+  int userId;
   int width;
   String description;
   String excerpt;
@@ -319,8 +305,8 @@ class ImageItem {
 
   ImageItem.fromParams(
       {this.height,
-      this.img_id,
-      this.user_id,
+      this.imgId,
+      this.userId,
       this.width,
       this.description,
       this.excerpt,
@@ -328,8 +314,8 @@ class ImageItem {
 
   ImageItem.fromJson(jsonRes) {
     height = jsonRes['height'];
-    img_id = jsonRes['img_id'];
-    user_id = jsonRes['user_id'];
+    imgId = jsonRes['img_id'];
+    userId = jsonRes['user_id'];
     width = jsonRes['width'];
     description = jsonRes['description'];
     excerpt = jsonRes['excerpt'];
@@ -338,6 +324,6 @@ class ImageItem {
 
   @override
   String toString() {
-    return '{"height": $height,"img_id": $img_id,"user_id": $user_id,"width": $width,"description": ${description != null ? '${json.encode(description)}' : 'null'},"excerpt": ${excerpt != null ? '${json.encode(excerpt)}' : 'null'},"title": ${title != null ? '${json.encode(title)}' : 'null'}}';
+    return '{"height": $height,"img_id": $imgId,"user_id": $userId,"width": $width,"description": ${description != null ? '${json.encode(description)}' : 'null'},"excerpt": ${excerpt != null ? '${json.encode(excerpt)}' : 'null'},"title": ${title != null ? '${json.encode(title)}' : 'null'}}';
   }
 }