+++ example/lib/photo_view_demo.dart
@@ -183,7 +183,7 @@ class _PhotoViewDemoState extends State<PhotoViewDemo> {
                                   ),
                                   LikeButton(
                                     size: 20.0,
-                                    isLiked: item.is_favorite,
+                                    isLiked: item.isFavorite,
                                     likeCount: item.favorites,
                                     countBuilder:
                                         (int count, bool isLiked, String text) {
@@ -264,12 +264,12 @@ class _PhotoViewDemoState extends State<PhotoViewDemo> {
     ///
     final Completer<bool> completer = new Completer<bool>();
     Timer(const Duration(milliseconds: 200), () {
-      item.is_favorite = !item.is_favorite;
+      item.isFavorite = !item.isFavorite;
       item.favorites =
-          item.is_favorite ? item.favorites + 1 : item.favorites - 1;
+          item.isFavorite ? item.favorites + 1 : item.favorites - 1;
 
       // if your request is failed,return null,
-      completer.complete(item.is_favorite);
+      completer.complete(item.isFavorite);
     });
     return completer.future;
   }